requirements.txt
django==4.2.21
psycopg2-binary==2.9.9 
python-dotenv==1.0.1
pillow
django-encrypted-model-fields==0.6.5
django-user-accounts


pybit==5.10.1
ta==0.10.2
pandas==2.2.3

celery==5.3.6
redis==5.0.8

matplotlib==3.9.2

asgiref>=3.3.2